Add workspace aliasing and use int in logger

+++ dpcpp/log/batch_logger.hpp
@@ -29,7 +29,7 @@ template <typename RealType>
 class SimpleFinalLogger final {
 public:
     using real_type = remove_complex<RealType>;
-    using idx_type = int64;
+    using idx_type = int;
 
     SimpleFinalLogger(real_type* const batch_residuals,
                       idx_type* const batch_iters)
