+++ reference/log/batch_logger.hpp
@@ -24,7 +24,7 @@ template <typename RealType>
 class SimpleFinalLogger final {
 public:
     using real_type = RealType;
-    using idx_type = int64;
+    using idx_type = int;
 
     /**
      * Constructor
